backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
aiohttp==3.13.5
beautifulsoup4==4.14.3
email-validator==2.3.0
